Centralize GP intake parsing
