+++ scripts/ingest_gp_files.py
@@ -9,106 +9,9 @@ import json
 from pathlib import Path
 from typing import Dict, List, Tuple
 
+from scripts.shared.gp_intake import extract_metadata_from_md, strip_base64_images
+
 ROOT = Path(__file__).resolve().parent.parent
-
-def extract_structured_sections(content: str) -> Dict[str, str]:
-    """Extract Insight/Gap/Fix sections from a GP Markdown snippet."""
-    section_re = re.compile(r'^\s*\*\s+\*\*The (Insight|Gap|Fix):\*\s*(.*)', re.IGNORECASE)
-    sections: Dict[str, List[str]] = {}
-    current: str | None = None
-
-    for line in content.splitlines():
-        heading_match = section_re.match(line)
-        if heading_match:
-            label = heading_match.group(1).lower()
-            sections[label] = [heading_match.group(2).strip()]
-            current = label
-            continue
-
-        if current:
-            if re.match(r'^\s*#{1,6}', line):
-                current = None
-                continue
-            stripped = line.strip()
-            if stripped.startswith('* *The '):
-                current = None
-                continue
-            if stripped:
-                cleaned = stripped.lstrip('> ').strip()
-                sections.setdefault(current, []).append(cleaned)
-
-    return {
-        label: ' '.join(lines).replace('\n', ' ').strip()
-        for label, lines in sections.items()
-    }
-
-
-def extract_metadata_from_md(filepath: str) -> Dict:
-    """Extract title, summary, and key sections from a markdown file.
-    
-    Addresses Protocol B: Fix Prompt Ingestion.
-    Prioritizes technical content (Insight/Gap/Fix) over conversational prompts.
-    """
-    with open(filepath, 'r', encoding='utf-8', errors='ignore') as f:
-        content = f.read()
-    
-    # Strip Base64 images early (Protocol D)
-    content = strip_base64_images(content)
-
-    # Extract first heading as title
-    title_match = re.search(r'^#{1,3}\s+(.+?)$', content, re.MULTILINE)
-    title = title_match.group(1).strip() if title_match else Path(filepath).stem
-
-    # Extract structured sections first (prioritize technical content)
-    sections = extract_structured_sections(content)
-    
-    # Build summary from structured sections if available, else use text
-    if sections.get('insight') or sections.get('gap') or sections.get('fix'):
-        # Use technical content as summary
-        summary_parts = []
-        if sections.get('insight'):
-            summary_parts.append(f"Insight: {sections['insight'][:80]}")
-        if sections.get('gap'):
-            summary_parts.append(f"Gap: {sections['gap'][:80]}")
-        if sections.get('fix'):
-            summary_parts.append(f"Fix: {sections['fix'][:80]}")
-        summary = ' | '.join(summary_parts)[:200]
-    else:
-        # Fallback: extract non-conversational text
-        clean_content = re.sub(r'[#\*_`\[\]]', '', content)
-        lines = [l.strip() for l in clean_content.split('\n') if l.strip()]
-        
-        # Skip lines ending with '?' (conversational prompts)
-        declarative_lines = [l for l in lines if not l.endswith('?')]
-        
-        # Filter out "Would you like" patterns
-        filtered_lines = [l for l in declarative_lines 
-                         if not re.search(r'would you like|shall I|do you want', l, re.IGNORECASE)]
-        
-        summary = ' '.join(filtered_lines[:3] if filtered_lines else lines[:3])[:150]
-
-    # Extract any keywords or key phrases
-    keywords = set()
-    patterns = [
-        r'(?:implementation|implement).*?(?:of|for)\s+([^.]+)',
-        r'(?:design|define)\s+([^.]+)',
-        r'(?:extends?|builds?|creates?)\s+([^.]+)',
-    ]
-
-    for pattern in patterns:
-        matches = re.findall(pattern, content, re.IGNORECASE)
-        keywords.update(m.strip()[:50] for m in matches if m.strip())
-
-    sections = extract_structured_sections(content)
-
-    return {
-        'title': title,
-        'summary': summary,
-        'keywords': list(keywords)[:3],
-        'insight': sections.get('insight', ''),
-        'gap': sections.get('gap', ''),
-        'fix': sections.get('fix', ''),
-    }
 
 def infer_target_module(content: str, title: str, keywords: List[str]) -> str:
     """Intelligently route GP content to appropriate Agda module.
@@ -157,18 +60,6 @@ def sanitize_string(value: str) -> str:
     escaped = value.replace('"', "'")
     escaped = escaped.replace('\\', '\\\\')
     return ' '.join(escaped.split()).strip()
-
-
-def strip_base64_images(text: str) -> str:
-    """Remove Base64-encoded images from text to prevent pollution.
-    
-    Addresses Protocol D: Purge the Noise.
-    """
-    # Strip data:image URIs
-    text = re.sub(r'data:image/[^;]+;base64,[A-Za-z0-9+/=]+', '[image removed]', text)
-    # Also strip inline image tags that might contain Base64
-    text = re.sub(r'!\[.*?\]\(data:image[^)]+\)', '[image removed]', text)
-    return text
 
 
 def generate_roadmap_step(gp_id: str, metadata: Dict, file_number: int, full_content: str) -> str:
